requirements.txt
flask
python-chess
gunicorn
Flask-Cors
